data: split group.js
