+++ src/data/things/group/index.js
@@ -1,0 +1,3 @@
+export * from './Group.js';
+export * from './GroupCategory.js';
+export * from './Series.js';
