+++ src/data/things/index.js
@@ -1,6 +1,7 @@
 // Not actually the entry point for #things - that's init.js in this folder.
 
 export * from './album/index.js';
+export * from './group/index.js';
 export * from './homepage-layout/index.js';
 
 export * from './AdditionalFile.js';
@@ -18,5 +19,4 @@ export * from './WikiInfo.js';
 
 export * from './content.js';
 export * from './flash.js';
-export * from './group.js';
 export * from './sorting-rule.js';
